송민규/일감유형완성/260622

+++ src/main/java/com/weple/cloud/system/service/TaskTypeVO.java
@@ -15,7 +15,7 @@ import lombok.ToString;
 @AllArgsConstructor
 public class TaskTypeVO {
 	private Integer typeId;
-	private Integer companyId;
+	private Long companyId;
 	private String  typeName;
 	private String  taskTypeDescribe;
 	private Integer position;
